+++ lib/capybara.rb
@@ -413,6 +413,8 @@ module Capybara
   require 'capybara/queries/title_query'
   require 'capybara/queries/current_path_query'
   require 'capybara/queries/match_query'
+  require 'capybara/queries/ancestor_query'
+  require 'capybara/queries/sibling_query'
   require 'capybara/query'
 
   require 'capybara/node/finders'